feature: ajustando estilos e formatações no front

+++ agenda-contatos-frontend/src/App.vue
@@ -45,6 +45,10 @@ const buscarTodos = async () => {
 };
 
 const excluirContato = async (contato) => {
+    if (!confirm(`Deseja realmente excluir o contato ${contato.nome}?`)) {
+        return;
+    }
+
     try {
         await axios.delete(`http://localhost:8080/contatos/excluir/${contato.id}`);
         buscarTodos();
@@ -74,27 +78,71 @@ onMounted(() => {
 
   <div :class="{ 'dark-bg': modalIsVisible }"></div>
   <main>
-      <header>
-        <h1>Contatos</h1>
-      </header>
-      <section class="grid-container">
-        <div class="button-right-container">
-          <button class="primary-button" @click="abrirModal">Adicionar Contato</button>
-        </div>
-        <MainGrid 
-          :contatos="contatos" 
-          @excluirContato="excluirContato" 
-          @editarContato="editarContato"
-        />
-      </section>
+      <div class="main-container">
+        <header>
+          <h1>Contatos</h1>
+        </header>
+        <section class="crud-container">
+          <div class="button-right-container">
+            <button class="base-button primary-button" @click="abrirModal">Adicionar Contato</button>
+          </div>
+
+          <div class="grid-title base-grid">
+            <div>Nome</div>
+            <div>Telefone</div>
+            <div>Cidade</div>
+            <div>Estado</div>
+            <div>Email</div>
+            <div>Categoria</div>
+            <div>Ações</div>
+          </div>
+
+          <div class="grid-container">
+            <MainGrid
+              :contatos="contatos"
+              @excluirContato="excluirContato"
+              @editarContato="editarContato"
+            />
+          </div>
+
+        </section>
+    </div>
   </main>
 </template>
 
 <style scoped>
 
+  main {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    background-color: #f0f0f5;
+    overflow: hidden;
+  }
+
+  .crud-container {
+    display: flex;
+    flex-direction: column;
+    flex: 1;
+    padding: 0 2rem 2rem 2rem;
+    min-height: 0;
+  }
+
+  .main-container {
+    height: 95vh;
+    width: 98vw;
+    display: flex;
+    flex-direction: column;
+    min-height: 0;
+    background-color: #f9f9fc;
+    border-radius: 8px;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+  }
+
   .grid-container {
-    height: 95%;
-    margin: 2rem;
+    flex: 1;
+    overflow-y: auto;
   }
 
   header {
@@ -102,13 +150,15 @@ onMounted(() => {
   }
 
   header h1 {
-  margin: 0.2rem 0;
-  font-weight: 500;
+    margin: 0.2rem 0;
+    font-weight: 500;
+    font-size: 4rem;
+    color: #333;
   }
 
   header p:first-child {
-  font-size: 3rem;
-  font-weight: 700;
+    font-size: 3rem;
+    font-weight: 700;
   }
   
   .dark-bg {
